+++ config/config.go
@@ -14,7 +14,7 @@ import (
 	"github.com/spf13/viper"
 )
 
-const VERSION = "26.10"
+const VERSION = "26.11"
 
 type Config struct {
 	Server              ServerConfig